[ISSUE-126] Move temperature to settings side bar
